lil refactor
