+++ cancer_ai/validator/model_runners/tensorflow_runner.py
@@ -1,0 +1,6 @@
+from . import BaseRunnerHandler
+from typing import List
+
+class TensorflowRunnerHandler(BaseRunnerHandler):
+    def run(self, pred_x: List) -> List:
+        return []
